requirements.txt
django
django_extensions
Pillow
python-dateutil

Faker